Deleted utils folder and moved classRoomLoader to ClassroomModal.js Removed some spacing I had
