--- client/src/pages/Dashboard/views/UserSettings/components/ClassroomModal/utils/useClassroomModalLoader.js
@@ -1,31 +1,0 @@
-import { useState, useEffect } from "react";
-import api from "utils/api";
-
-export const useClassroomModalLoader = (roomId) => {
-	const [room, setRoom] = useState();
-
-	useEffect(() => {
-		if (false === roomId) {
-			setRoom(null);
-			return;
-		}
-
-		if (!roomId) {
-			setRoom({ name: "" });
-		} else {
-			const getRoom = async () =>
-				setRoom((await api.getClassroom(roomId)).data);
-
-			try {
-				getRoom();
-			} catch (err) {
-				// TODO error handling
-				console.log(err);
-			}
-		}
-
-		return () => setRoom(null);
-	}, [roomId, setRoom]);
-
-	return room;
-};
